+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Parameter handling functions : attList, attMap, attStr
+
 ## [0.2.0] - 2024-05-09
 
 ### Added